Add sleeps to tests that rapidly alter NFS config
We have some CI tests that alter running NFS server configuraiton
and verify that the results were applied successfully. Depending
on situation this can bump into systemd limits for how rapidly
services are allowed to restart. Introducing 5 second sleeps
between tests where we do this ensures that we don't bump up
against the limit.

The alternative course of changing systemd defaults for the
testing runs was decided against because we want to mirror
actual systems as closely as possible. Refactoring tests
to avoid triggering the issue ended up being rather impactful
and so simple fix was adopted.