Add handling for MatchNotFound to share_info parsing
There are some situations where domain SID for server may change
due to administrator action, which causes existing SID entries
in share_info.tdb to no longer resolve. Due to switching to
python bindings for this, SID lookups can fail with MatchNotFound
and so add an additional error message / case for this.